+++ examples/dataflow-hashpipeline/Dockerfile
@@ -4,11 +4,12 @@ ARG WORKDIR=/dataflow/template
 RUN mkdir -p ${WORKDIR}
 WORKDIR ${WORKDIR}
 
-COPY requirements.txt .
-COPY hashpipeline.py .
+COPY hashpipeline ./hashpipeline
+COPY main.py .
+COPY setup.py .
 
-ENV FLEX_TEMPLATE_PYTHON_REQUIREMENTS_FILE="${WORKDIR}/requirements.txt"
-ENV FLEX_TEMPLATE_PYTHON_PY_FILE="${WORKDIR}/hashpipeline.py"
+ENV FLEX_TEMPLATE_PYTHON_SETUP_FILE="${WORKDIR}/setup.py"
+ENV FLEX_TEMPLATE_PYTHON_PY_FILE="${WORKDIR}/main.py"
 
 RUN pip install -U apache-beam==2.21.0
-RUN pip install -U -r ./requirements.txt
+RUN python setup.py install
